new update -- working in cogs
